[#239] test: CommentLikeControllerTest 생성 및 테스트 코드 이동
